pyproject.toml
[project]
name = "knowledge-assistant"
version = "1.0.0"
description = "MCP Server for Knowledge Vault exploration and Q&A"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "structlog>=24.0",
    "aiofiles>=23.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
dev = [
    "ruff>=0.2",
    "mypy>=1.8",
    "types-PyYAML",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
knowledge-assistant = "src.server:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
